# swagger.yaml
openapi: 3.0.0
info:
  title: Artists · MEN REST API
  description: MongoDB ExpressJS NodeJS REST API
  version: 1.0.0
servers:
  - url: https://artists-men-restful-api.herokuapp.com/api/
    description: Heroku remote server
  - url: http://localhost:4000/api/
    description: Localhost development server
components:
  securitySchemes:
    ApiKeyAuth: # arbitrary name for the security scheme
      type: apiKey
      in: header # can be "header", "query" or "cookie"
      name: auth-token

  schemas:
    Artist:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        info:
          type: string
        genre:
          type: string
        photo:
          type: string
        listeners:
          type: string
        albums:
          type: array
          items:
            type: string
        top_tracks:
          type: array
          items:
            type: string
        similar_to:
          type: array
          items:
            type: string

    User:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        date:
          type: string

paths:
  /user/register:
    post:
      summary: Register a new User
      tags:
        - Users
      description: Register a new User in the database
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"

      responses:
        201:
          description: User created succesfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  _id:
                    type: string

  /user/login:
    post:
      summary: Login a User
      tags:
        - Users
      description: Login a User in order to get JWT token (auth-token)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string

      responses:
        200:
          description: User logged in succesfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  data:
                    type: string

  /artists:
    get:
      summary: Retrieve all Artists
      tags:
        - Artists
      description: Retrieves a list of all the Artists as JSON objects.
      responses:
        200:
          description: A list of artist JSON objects in an array.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Artist"

    post:
      summary: Create a new Artist
      tags:
        - Artists
      description: Creates a new Artist
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Artist"

      responses:
        201:
          description: Artist created succesfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artist"

  /artists/{id}:
    get:
      summary: Retrieve Artist
      tags:
        - Artists
      description: Retrieves a specific Artist based on its id.
      parameters:
        - in: path
          name: id
          required: true
          description: Artist Id
          schema:
            type: string
      responses:
        200:
          description: An Artist in the format of a JSON object.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Artist"

    put:
      summary: Update Artist
      tags:
        - Artists
      description: Updates a specific Artist based on its id
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: id
          required: true
          description: Artist id
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Artist"

      responses:
        201:
          description: Artist updated succesfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artist"

    delete:
      summary: Delete Artist
      tags:
        - Artists
      description: Deletes a specific Artist based on its id
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: id
          required: true
          description: Artist id
          schema:
            type: string

      responses:
        201:
          description: Artist deleted succesfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artist"